rbbz: ensure cookies are stored per instance (bug 1139214); r=smacleod
As I was working on Selenium tests, I noticed a peculiar issue: running
multiple Selenium tests back-to-back from the same nose/Python process
was producing an error around Bugzilla tokens being invalid.

After a few hours chasing false leads, I eventually captured a packet
trace and noticed something weird: xmlrpc requests contained multiple
Bugzilla cookie values!

I set some breakpoints, and sure enough the "cookies" attribute of the
transport class was effectively stacking entries. This appeared to "just
work" in many scenarios. However, I appeared to jiggle it in just the
right way to cause a failure.

This patch fixes the problem by ensuring that the "cookies" attribute
is an instance attribute, not a class attribute. This should be inside
an __init__. However, xmlrpc.Transport is an old style class and there
is only pain and suffering if we introduce an __init__ here (due to some
multi-inheritance later on in the class hierarchy). Using @property is a
bit hacky, but it solves the problem.

With this patch applied, multiple Selenium tests magically started
working again.