qimportbz: allow sorting the patch list by patch description (bug 1001463)
Currently the patch list is sorted only by id. This adds a new option 's'
which toggles the sort order between attachment id and attachment description.

In the "all patches are wanted, but they are in the wrong order" case, this
means instead of having to type out various patch numbers/ranges, a simple 's'
<enter> <enter> will work. In the "we only want some of the patches" case, this
still improves UX, since it's more likely the desired patches are in a
contiguous sequence, and so a range can be used instead or multiple separate
patch numbers.