pash: store user in a variable; r=bkero
os.getenv('USER') was used multiple times in this file. The content of
the environment variable doesn't change during execution. So get the
value once and store it in a variable.

Also, os.environ is the preferred mechanism to access environment
variables from Python. os.getenv is a low-level API into the POSIX
function of the same name. The difference doesn't matter in this use
case, so use the method that is more accepted by the Python community.