pash: move non-root login code path to isolated function (bug 1140214); r=bkero
We want the root login code path to be clean so that things like import
errors don't result in root login being denied. This commit moves the
non-root login code path to a separate function. We also move imports
into this function so they aren't evaluated until a non-root login is
performed.

With this commit, the root login code path is now only requires a
pash.py file with proper Python syntax and a working Python installation
(since all we import are modules from the core Python distribution).