autoland: create .t style tests for autoland (bug 1140155); r=gps
This adds mach commands for interacting with autoland and replaces the current
unittest based integration test with a .t style test which uses these commands.

This also fixes bad paths in the autoland.conf and start-autoland.sh commands
which were introduced when the volume mapping was being used for debugging.