testing: move make_admin into MozReviewBoard class; r=smacleod
We want to call this API from elsewhere. Move it out of an inline mach
command and into a module.

As part of this, we changed the _get_rb() function to take an optional
path argument. It is a holdover from the days of old to pass a path into
the reviewboard mach commands. This paves the road to getting rid of
this pattern.