mozreview: print submitter username in dumpreview output; r=smacleod
Managing the active user in MozReview tests is a bit overwhelming. It
can result in unexpected users being used.

To help avoid confusion, this patch adds the submitter the the
"dumpreview" command output.

The new test output clearly demonstrates that we have the admin user
performing a lot of actions. Fortunately, the admin user isn't actually
a Review Board admin currently. But we want to change that. Upcoming
patches will change these tests to use a different user and the new
output will prove the admin user is no longer being used by tests.