mozreview: create and use adminbugzilla alias; r=smacleod
We have a goal to make the default admin user an actual admin in Review
Board. Before this can be done, we need to change tests to stop using
this user.

To make the transition easier, we introduce an "adminbugzilla" alias
that runs the Bugzilla helper command with the admin credentials. We
switch invocations of "bugzilla" requiring admin access to this new
alias.