mozreview: convert test-auth.t to have default user; r=smacleod
We have a goal to introduce a default, regular user to the common
environment. test-auth.t is one of the more complicated tests involving
users. Convert it to the new format first. As part of this, we also make
the admin user an actual admin sooner, as this will also soon be done in
commonenv.