mozreview: create a default user as part of commonenv; r=smacleod
Before we can make the admin user an actual admin user, we need to have
tests stop using the admin user as part of regular Review Board
interaction. This patch moves the creation of the default user into
commonenv and updates all remaining tests to reflect the utilization of
this new account.