bzexport: don't use admin user in tests; r=smacleod
bzexport tests, like mozreview's, were using the admin user as part of
testing. That's a bad practice. This patch puts an end to the practice.

The test output did change in one unexpected way. The bug status is now
marked as ASSIGNED instead of NEW in a few locations. This appears to be
purely a result of the differences between admin and non-admin users.
But the change does warrant additional scrutiny in case it isn't.