hgmo: make replication work
The SSH connectivity from mirrors to master wasn't working for several
reasons. Bad permissions on SSH keys and the known hosts file. The
vcs-sync@mozilla.com account not being defined in LDAP. This patch fixes
all of these.

With this patch, pushing to the SSH master should now mirror properly
\o/