autoland: add mach command to wait for pingback
This adds a mach command which waits for a specified period of time for a
pingback from autoland. Using this command in tests will likely have to wait
until reviewboard is running in a docker image as communication back to the
docker host is problematic under boot2docker.

This also fixes the tests which had the order of the source and destination
trees reversed, and improves the handling of an unknown destination tree in
autoland.