hgmo: collect Python code coverage from inside Docker containers
Code coverage is a wonderful thing. Unfortunately, Docker containers
make collecting code coverage difficult for a few reasons:

1) Code coverage is typically something you have to explicitly enable at
   process start time.
2) Aggregating coverage results from files inside the container to files
   under version control is difficult due to filename mismatch.

This patch aims to solve these problems.

We install a custom "sitecustomize.py" file in the Docker container.
This file gets loaded every time the Python interpreter starts. Our
implementation of this file looks for a "signal" to perform code
coverage. If it is present, coverage is collected for the current Python
process automatically and the results are written to a randomly
generated filename at a well-defined location. This all happens
automatically: just configure the container once and coverage will be
collected.

To support retrieving code coverage results, we introduce an API to
fetch the contents of a directory from a running Docker container. We do
this by invoking `tar` via an `execute` Docker API call and feeding the
results into Python's `tarfile` API. This is a bit hacky, but it works.

We use the directory-retrieving API to obtain all code coverage files
from the container. These files are pickled data structures. We unpickle
them and convert them back into the data structures expected by the
"coverage" Python package. As part of this deserializing, we remap
filenames inside the container to files on the host.

The final step is to save the collected and normalized code coverage
results files to our aggregation directory on the host. From there, the
existing code in run-mercurial-tests.py takes over and incorporates them
as part of the generated code coverage report.

The implementation likely has minor bugs. There are certainly no
shortage of improvements that could be made. The manual filename mapping
is particularly annoying. But perfect is the enemy of done. This
solution appears to work well enough. We now have code coverage from the
custom shell installed on the sshd server inside the hgmaster container
(pash). It is already telling me useful things about what tests to write
next.

The approach for collecting coverage could be applied to other
containers. I had an immediate need to collect pash's coverage, which is
why I started there.