hgmo: add a special syntax to Dockerfiles to declare context files
Previously, we maintained a special list in code for declaring which
files from source control were added to the Docker build context. Having
this data live in code was not ideal because it was separated from where
it was used. The location wasn't ideal.

We introduce a special syntax for declaring extra context files to
Dockerfiles themselves. The declaration now lives next to its users and
all is well again with the world.

The next best thing would be to parse ADD directives. However, that
involves glob matching and other nastiness that I have no inclination to
support at this juncture. This syntax is simple and gets the job done.