docker: limit the set of files that cause bmodb and bmoweb image invalidation
We were including every file from v-c-t in the bmodb and bmoweb images.
This was causing them to get invalidated (and requiring a new
bootstrapped image) any time a file in v-c-t changed. This was extremely
wasteful.

This patch limits the set of files whose changes can cause the bmo
images to be invalidated.