mozreview: use an explicit and temporary file for cookie storage (bug 1115045); r=smacleod
RBTools insists on creating a cookie file. Without this hack, it will
try (and fail) to create a cookie file in the user's home directory
(which likely doesn't exist in production) or in / (if the home
directory could not be resolved - this also happens in production).

This patch uses a temporary file for cookie storage. It should prevent
permission errors in production.