autoland: initial support for landing github patches (bug 1148297)
This adds a pullrequest table to the autoland database and a REST endpoint for
request pullrequests to be landed to the mozreview repo. Autoland will
periodically check for pending pullrequest to be landed and attempt to land
them, updating github with the results.

We also remove the pingback_url from the status endpoint in case that is
sensitive information.