testing: relicense run-mercurial-tests.py to MPL
It was GPL because it was calling into GPL'd Mercurial functions. We no
longer call into Mercurial code, so there is no requirement for the GPL.

I own blame for every line in this file. As the sole author, I believe
it is within my right to relicense this code as I see fit. And I choose
MPL, as that is the default for all code in this repo.