testing: create and use a context manager for auto cleaning Docker orphans
If a process that starts Docker containers is aborted, the Docker
containers may become orphaned. When the test harness is aborted, dozens
of containers may become orphaned.

While tests should do a better job at cleaning up after themselves, this
isn't achievable in today's Mercurial test harness. So, we invent a
context manager that auto cleans unknown containers and images when it
exists. We hook this context manager up to run-tests.