testing: kill wsgi processes when new repos created
test-unicode.t was failing due to a 404 on a file revision lookup. What
was strange was that this exact same URL returned 200 just a few moments
earlier!

It appears our hacky "refreshinghgwebdir" extension didn't work. Killing
the WSGI processes to force a new scan does work. It's hacky. But it
works.