Bug 1055021 - mozreview: Get all review-request summaries for a given bug. r=smacleod
This adds a get_list() handler to the ReviewRequestSummaryResource, which
can be filtered by bug.  This should also make the get() call slightly
more efficient.

This feature is required by the Bugzilla MozReview extension, since we won't
be posting attachments for parents anymore, and eventually we will also
allow different bugs in one review-request "family", but we'll want to
display the entire family on every associated bug, for completeness.