hgext: use public field rather than status field to indicate drafts (bug 1144085) r=gps
We're currently using the review status field to indicate to the user whether
or not a review has been published, but this is incorrect. This patch adds the
public field to the reviewdata and updates the client to use this instead. It
would be nice to also remove the status field from the data presented to the
client, but I think this will create more confusion than good until we have
a means of indicating to people that they should update their clients.