ansible: support locking in stingray_node module
When manipulating the state of nodes in Stringray, it appears that the
only way to change state is to pass in lists of nodes and their expected
state. This, of course, is prone to races between multiple clients
simultaneously updating pool state. 2 clients could each retrieve the
state of the pool and then each perform their own update. Last write
wins. This concern is not theoretical: I witnessed a race in Ansible
when marking multiple nodes as drained.

To eliminate this race, we add file locking support to Stingray pool
manipulation. If the Ansible playbook specifies a lock file, we acquire
a global lock around the pool manipulation to ensure only 1 local client
can modify the pool at once.