hghooks: no-op during pulls (bug 1159099); r=bkero
hgweb nodes are mirrors of the master server. They ideally run the same
hgrc files, hook configurations and all.

Today, hgweb nodes have a directory of hooks (library_overrides) that
are modified versions of the hooks that simply no-op. Maintaining this
separate list of overrides is tedious and prone to issues. For example,
when someone introduces a new hook, one must also create an override for
that hook or else the hook will run during `hg pull` operations on the
mirror. In many cases, hooks running as part of `hg pull` (or even `hg
clone`) will abort the operation when the master didn't. This can lead
to mirrors getting out of sync with the master.

This patch modifies existing hooks to take the "source" of the
changegroup into consideration. If the changegroup is coming from a
"pull", we now return 0.

Previously, many hooks were a no-op for "strip" events. We've extended
functionality to encompass "pull" as well. Some hooks were previously
not ignoring strips. We've modified these hooks to be consistent with
every other hook.