testing: support refreshing bmoweb container (bug 1160029)
We want the ability to restart httpd in bmoweb so local code changes are
picked up. With mod_perl, the parent process is loading some code, so we
need to perform a hard restart of the entire httpd process tree to pick up
changes. Docker containers terminate when pid 1 exits. So we need to
introduce a new pid 1. Add supervisor to fulfill that role.

While we are here, we also hook bmoweb up to `mozreview refresh` to
restart the httpd service.