testing: don't use supervisor in bmoweb
This is a partial backout of 3e2d3a7ccd73 (bug 1160029).

For reasons that completely confound me, this patch is breaking
Bugzilla.

When Apache is run under supervisor, `mozreview create-user` is failing.
Under the hood, it establishes a connection to Bugzilla's XML-RPC
interface. It calls User.login() and User.create(). The first request
succeeds but the second one yields an empty response from the server.
And by empty, it looks like the connection just drops after the 2nd HTTP
request is sent across the wire. And there is nothing in the logs on the
server to indicate any kind of failure. Dropping supervisor and running
/run-apache.sh directly makes the problem go away and restores working
behavior.

When I examine both versions of the container, the process tree looks
identical. We have apache2 running as root with a few child processes
running as the "bugzilla" user. What really boggles me is that the first
HTTP request on the TCP connection succeeds but the second one always
reliably fails. I could understand if the first request fails. But to
have the second request on a (premably keep-alive enabled connection)
dropping is most weird.

This bug defies reason. Until I have time to investigate, disable
supervisor so people's dev environments aren't broken.