MozReview: Hide draft banner buttons on commit requests (Bug 1162111); r=mcote r=mdoglio
Assign css classes "parent-request" or "commit-request" to the review
request div depending on the type of request. These classes can be used
to apply CSS to only a single type of request.

Hide the draft banner buttons on all requests and unhide them if we are
viewing a parent request.