mozreview: refactor code for determining what to review; r=smacleod
The code for determining what to review was previously spread out over a
few functions and was difficult to comprehend and change.

We introduce a single function (which runs as part of discovery) that
determines which changesets are to be reviewed and it saves that list
inside the "pushoperation" instance. This allows all subsequent
consumers to easily see what is going to be reviewed. We plan to
leverage this work in a subsequent change to do more processing at push
time.

Functionality is mostly identical. Some refactoring has been performed
to clean up code that doesn't run anymore. This is most notable with
the usage of "pushop.outgoing.revs" to determine review nodes: this
never occurs because either the user specifies "-r" or "-r ." is
implied. This wasn't obvious before. But post code refactor it is pretty
clear.

Review-URL: https://reviewboard.mozilla.org/r/8383