testing: log in newly-created Bugzilla users (bug 1164688); r=smacleod
Bugzilla recently made it so users with no activity in the past 60 days
can't have review flags aimed at them. See bug 751862.

We change the user creation code to log in the user by default to
simulate Bugzilla activity. This is easier than modifying every test to
touch newly-created users.

This required a bit of refactoring to make the login code reusable.