hghooks: use env in shebang (bug 1166065); r=bkero
We're transitioning to run Mercurial from Python 2.7. If we hardcode
/usr/bin/python, we'll get the system Python, which on CentOS 6 machines
is Python 2.6. Use /usr/bin/env to find the appropriate Python
interpreter for the environment.

Of course, hooks should all be using the "python:" prefix so they are
loaded as Python modules and we don't incur avoidable overhead running
new processes.