autoland: create separate table for mozreview updates
Right now we're keeping track of whether or not data is posted to mozreview
inside the Transplant table. Since we'll also need to do this for pull
requests, it makes more sense to keep a separate table of data to be posted
and do things all at once.

Since we're here, we also add support for mozreview updates when importing
pullrequests.