bundleclone: support client-side preferences for choosing bundles (bug 1041173); r=smacleod
The bundle manifest supports defining arbitrary key-value metadata for
each entry. The original intent of this metadata was to facilitate
"content negotiation" to allow clients to select the bundle that is best
for them. This commit implements that feature.

We implement support for the client to define a list of preferences for
attribute values. The implementation is about the simplest possible. It
is far from a robust content negotiation mechanism. But it should be
good enough for now. If needed, we can always implement allow or deny
filters to prune the manifest so undesired fallback doesn't occur. Or,
we can implement a query mini language. There are infinite
possibilities. But KISS applies.

New tests covering the feature have been implemented. Since Mozilla's
bundles will likely only vary by compression format and EC2 region, we
use these attributes in our tests.