bundleclone: copy stream clone handling code from Mercurial; r=smacleod
Streaming clones are significantly faster than changegroup-based clones.
Using them will result in significant wall time wins when cloning.

To support streaming clones, we need to copy some code from Mercurial
for dealing with them. This is because the existing code is buried in
closures and/or tightly coupled to the wire protocol. I have already
sent patches upstream to decouple the code. But, the damage is done:
since we want to support Mercurial clients <3.5, we're forced to copy
the code.

This review should largely be a rubber stamp, since this code was
effectively copied from upstream.