bundleclone: support stream bundles (bug 1041173); r=smacleod
`hg clone --uncompressed` engages a special clone mode where files from
Mercurial's store are effectively streamed from the server. It is
roughly equivalent to piping a tar file over netcat. This mode consumes very
little CPU on both the server and client, as both ends are essentially doing
little more than raw I/O.

On mozilla-central, a regular clone (equivalent to an unbundle of a gzip
bundle file) consumes ~340s CPU time on the client. With streaming
clones, that number is under 60s, with over half that time spent in the
kernel. Wall time goes from ~360s to ~90s, a 4x reduction. (Measurements
obtained on OS X with a 2014 MBP.)

The faster speed does come at a cost: transfer size. Streaming clones
are larger than bundle transfer based clones. On mozilla-central, the
bundle size increases from ~1024MB to 1480MB, an increase of 45%. For
many network connections, the increased size results in a long transfer
time than CPU saved. However, for high speed networks (such as inside
EC2 where a 1Gbps link is available), the extra data can be transferred
with just a few seconds of overhead and overall clone time is reduced.

Unfortunately, Mercurial doesn't yet provide a way to generate or apply
these "stream bundles." Instead, this functionality is tightly coupled
with the wire protocol.

This patch implements support for producing and consuming "stream
bundle" files. If a server advertises a stream bundle, the client will
use stream mode to apply it.

When this is used in Firefox automation, clone times should be reduced.

When this is deployed to Firefox OS automation, they can likely delete
all their code doing repository management via tar files, as this is
essentially the same thing. Even if this solution is a few seconds
slower than tar files, it is a better solution because it is built on
top of Mercurial primitives and thus should be aware of any special
considerations that a "dumb" tool like `tar` may not be. It is thus a
more robust solution.

I've spoken with Mercurial maintainers and they are keen with this
feature being part of Mercurial core. We could probably have it
implemented in Mercurial 3.5. Once that is deployed, the bundleclone
extension becomes obsolete.