mozreview: add import page for github pullrequests (bug 1148297); r=smacleod,r=mdoglio
This adds backend support for importing pull requests into mozreview. Similar
to autoland to try, we create a table that allows us to track the originating
github user, repo and pull request and the resulting review request id. An
initial version of the UI is present to allow the backend work to be tested.
It will be revisited once work to allow autofiling of bugs is completed.