reviewboard: we should not reset reviewers if none are specified in commit summary (bug 1169499); r=gps
This is a symptom of a larger problem where specifying reviewers in the commit
summary causes any reviewers specified in the UI to be lost. Solving that
larger problem requires us to have a good means of identifying reviewers which
were specified in the UI and which were specified in the commit summary.

In the mean time, if no reviewers are specified in the commit summary we can
skip setting the reviewers on the draft, which will preserve reviewers that
were set in the UI if none are specified in the commit.