ansible: install CRON to periodically generate repository bundles (bug 1041173); r=bkero
Now that we support generating Mercurial bundles and uploading them to
S3, let's make generating them part of our standard operating procedure.

This commit installs daily CRON entries to do the bundle generation.

We introduce the "outputif" script to execute a process and only print
its output if it exits with an error. This makes CRON email tolerable by
not sending it if everything is OK.