scripts: don't chown bundle manifest file; r=bkero
The hg user isn't a member of the scm_* groups, so it can't change group
ownership. This isn't strictly required for bundle generation to work.
So we disable it. We leave the comment in because this is something
we'll want to fix someday, as all files in .hg should really be owned by
the same group.