mozreview: generate commit IDs when pushing to review repo (bug 1160266); r=smacleod
Previously, we only generated commit IDs at commit time if the
repository was known to be associated with a review repository. Now that
we have support for mass rewriting changesets, start generating commit
IDs at push-to-review-repo time.

Since we now add commit IDs to every to-be-reviewed changesets, all
SHA-1s have been updated.

Review-URL: https://reviewboard.mozilla.org/r/8395