mozreview: abort push when working directory is dirty; r=smacleod
Upcoming patches will perform more commit rewriting after push. We don't
want to get in a scenario where the push and/or review request succeeds
but the client can't update commits, as that would result in metadata
getting out of sync. With this change, we verify rewrites can be
performed *before* push and fail fast otherwise.

Review-URL: https://reviewboard.mozilla.org/r/8447