mozhg: implement changeset rewriting API; r=smacleod, dminor
Mass modifying changesets and all their children is non-trivial because
there isn't an API in core Mercurial for this, at least not one that is
available to the versions of Mercurial we need to support. In Mercurial
3.4, histedit is kinda/sorta extensible and may work for this purpose.
In the future, apparently lots of histedit's functionality is going to
get rolled into core to make this easier. In the mean time, we need
to implementat functionality ourselves.

This commit introduces a generic changeset rewriting API. Pass in a list
of nodes and a function to produce a new commit and it takes care of
phases, bookmarks, obsolescence, stripping, working directory updates,
and everything else you want it to do.

Basic tests have been included.

Review-URL: https://reviewboard.mozilla.org/r/8281