mozhg: properly compute status when rewriting impacted changesets
We weren't computing status properly. This resulted in file
modifications being dropped and empty changesets being created.

This is a pretty nasty bug and could lead to lots of pain for
developers. Pushing it without review since it is 1830 on a Friday and
nobody is around to grant review.

It's worth noting that similar code exists in the tests. The testing
code is computing status properly. This code isn't.