autoland: add capability to autofile bug if no bug id is specified
If no bug id is specified for a pullrequest to be landed, first the pullrequest
title is examined for a bug id. If none is found there, a new bug is filed
automatically using the pull request title and description for the bug summary
and first comment, and default values for product, component and version.