pushlog: render pushlog info in more places in hgweb (bug 1171321); r=smacleod
I added a feature to Mercurial 3.4 that allows you to easily monkeypatch
the functions that provide per-changeset data to the templater for hgweb
pages. Now that we have Mercurial 3.4.1 deployed on hg.mozilla.org, we can
start using this feature for awesomeness.

This commit exposes pushlog metadata to the templater and renders it on
per-changeset pages and the changelog template. When looking at these
pages, you now don't have to click another link to find pushlog info.