hgmo: render reviewer and bug list in hgweb (bug 1171321); r=smacleod
Let's display more useful info on per-changeset pages on hgweb!

This commit introduces support for rendering the list of reviewers and
bugs.

The state of the hgweb extensions on hg.mozilla.org is kinda funky. Most
of the code today lives in the poorly named hgext/pushlog-legacy
directory (named this way because content came from the original
pushlog repository that was imported into v-c-t around the same time
work started on the new pushlog extension, which is under hgext/pushlog).
So, given that the code is in a weird state and I'm about to start work
to pour a lot of love into it, I took the opportunity to create an
"hgmo" extension. My hope is this extension will become the location for
all the hgweb customizations we do on hg.mozilla.org.

This extension relies on a v-c-t checkout, which we currently don't
have on the hgweb machines. We'll need a follow-up commit with Ansible
changes to teach the hgweb machines how to run from v-c-t. Since
the hgrc for hg.mozilla.org is stored in this repo and we haven't
introduced this extension to it yet, this commit is safe to check in.