hgmo: show that a commit was backed out (bug 1171321); r=smacleod
When looking at commits, it would be nice if we could quickly see if the
commit was backed out. This commit implements support for that feature.

"Bad" commits are denoted two ways. First, a pile of poo is rendered
next to the changeset node in the page header. Second, a "backed out by"
message is rendered at the top of the changeset metadata section. It's
hard to miss.

The styling isn't the greatest. But something is better than nothing.