hgmo: add --hgmo flag to hg serve (bug 1172677); r=smacleod
With this commit, one can run `hg serve --hgmo` to launch an hg serve
process with a configuration very similar to hg.mozilla.org. This makes
it super easy to hack on templates.

We add some documentation for how to hack on the theming so we have
something to point people at if they want to contribute.