mozreview: Create model for storing MozReview user information (Bug 1160517). r=mcote
Having a MozReviewUserProfile models allows MozReview to tie extra
information to a Review Board user. Before this commit that has been
done using a separate table for the specific information (such as the
BugzillaUserMap).

In order to support querying ldap for user permissions an ldap username
must be stored for each Review Board user. Rather than make a specific
table for this, introduce a generic MozReviewUserProfile and store the
information there.