Bug 1173441 - Optimize pushhead(tree) revset. r=gps
This patch adds an index for the specific query, guards against bad
pushlog data, and makes the revset lazy.  Locally, it performs
significantly better than the earlier version.  pushhead() without an
argument is not performant enough to use locally without this patch;
with this patch, it is usable.

My use case is finding pushes "close" to tip.  Limiting the query to
pushes in a certain time range is probably the way to improve the
performance.