mozreview: Add mach command for dumping ldap association (Bug 1160517). r=gps r=dminor
In order to test how interactions influence a users ldap association we
need a way to query it. Add a command to print the ldap username
associated with a MozReview user.

We always use a hardcoded username/password when querying for ldap
credentials as only admins or users with a special permission may
retrieve ldap usernames for users besides themselves.