mozreview: Add function to update ldap association (Bug 1160517). r=gps r=dminor
Add `associate_ldap_username` function to pushhooks.py which will update
a users associated ldap_username through the web API. This is a new
function rather than a modification to post_reviews so that we could
theoretically introduce a new wireprotocommand dedicated to associating
your ldap account (which could be used by non hg developers, rather
than having to push actual code). To start we'll probably just use it
when submitting a review.