testing: detect stopped Docker containers when waiting for servers (bug 1173880)
If a Docker container doesn't initialize properly, we'll spin for many
seconds (possibly minutes) waiting for services to come online. All this
does is waste people's time waiting for a service to start that will
never start.

This commits changes the service polling to also check whether the
underlying Docker containers are still running. If the container stops,
we stop polling immediately.