autoland: retry autoland requests if the destination tree is closed (bug 1174256) r=mdoglio
This adds code to check tree status and not attempt to push to a tree if it is
reported as closed. We also add a check for 'is CLOSED' in the response from
the mercurial server and will not update the status of the autoland request in
that case so that it will be retried next time.

This also fixes a problem with the autoland docker packaging and the error
handling when autoland can not connect to mozreview that were noticed while
testing.