ansible: extract MozReview virtualenv creation to own role (bug 1176044); r=smacleod
This functionality is shared by the admin server, dev server, and
Docker. But it is mostly duplicated 3 times. Establish a shared role to
hold the logic. We base it on the reviewboard-admin copy of the code
because it is the most well implemented. Since we took the
reviewboard-admin code, refactor it to use the new role.