mozreview: Stop storing Bugzilla cookies (bug 993233). r=smacleod
Since all communication is now done via API keys, we no longer need to
store cookies in the user object.

For the purposes of identification with MozReview, logging in with username
and password or login cookies is still possible via the API and with
/account/login/?full=true (or any other value for "full"). The latter is
intended only for the admin account, as it will not cause an API key to be
generated and sent over from Bugzilla, meaning most calls from Review Board
to Bugzilla will fail. In both cases, the cookie/password is only used to
verify the account with Bugzilla and is discarded immediately after.