autoland: remove testrun monitoring code (bug 1181139) r=mdoglio
This removes the code to monitor test runs. Our plans for autoland to inbound
are to only verify that a try run has been made, not check for failures.
Presumably at some point we'll want something more sophisticated in place for
this, but I think that it makes more sense to add that to the MozReview UI as
it makes it easier to provide feedback to the user as to why their revision
can't be autolanded and what steps they should take to fix it.