reviewboard: properly send error responses; r=smacleod
wireproto.pusherr was being used to communicate errors to the client. The
way this works under the hood is that a literal string is sent to the
client, which is expected to infer that a string (as opposed to say an
integer in string form) is an error message. This completely confuses
the client-side parser, as it expects a newline based protocol.

Since pusherr is only really appropriate for errors during the push
operation, we leverage a custom type that has the same effect. It has
built-in knowledge of the wire protocol and automatically serializes to
use the "error" line command to represent the error.

Sadly, we have no test coverage for this. But an upcoming commit will
add some.