reviewboard: advertise features in mozreview capability; r=smacleod
Adding separate capabilities for each reviewboard feature isn't the
preferred way to do things. Use a list of features in the "mozreview"
capability instead.

By committing only changes to the server, we prove that old clients are
compatible with the new server code.

Unfortunately, we can't easily repurpose the "reviewboard" capability
for the features list because client.py's auto redirect code effectively
does "if not caps['reviewboard'] and caps['listreviewrepos']" meaning
that adding "listreviewrepos" to the "reviewboard" capability wouldn't
satisfy the "not reviewboard" check and would cause auto discovery to
fail. Once sufficient time has passed and all clients in the wild have
been upgraded, we can consider removing this advertisement at the risk
of breaking old, un-upgraded clients.