reviewboard: look for capabilities in mozreview capability; r=smacleod
Now that the server exposes a comma delimited list of features in the
"mozreview" capability, switch the client to use it.

Previously, presence of the "reviewboard" capability implied the ability
to push reviews. The equivalent is now the "pushreview" sub-capability.

This code is not backwards compatible with old servers. We should not
land this change until the server starts advertising the "mozreview"
capability.