reviewboard: remove duplicate code around review request updating; r=smacleod
Before this commit, the code for creating a new review request and
updating an existing review request was 95% similar. The only difference
was the function was storing the commit id and the in-line code
was populating it at review request creation time. In the end, it didn't
matter since the property was being assigned to the review request
draft.

The return value from update_review_request has been dropped since rr is
now passed in.