reviewboard: return review data as dicts, capture reviewers; r=smacleod
Before this commit, classes from the RBTools package were being returned
from reviewboardmods.pushhooks.post_reviews() and operated on inside the
Mercurial extension. This was arguably a layering violation. In
addition, since we don't import the rbtools modules from the Mercurial
extension, it meant we couldn't do things like catch the APIError from
the extension, which is needed for looking up the draft review request.
By returning dicts instead of RBTools classes, we work around both
deficiencies.

While we were here, reviewers are now returned from this function. This
will be used in a subsequent commit.