reviewboard: consider draft review requests for public state; r=smacleod
As part of developing an unrelated feature, I discovered that the review
request "public" field sent to clients was wrong in the scenario where a
push had occurred to an existing review request. Essentially, the
"public" value was coming from the review request. If a review request
had never been published, it was always not public. If it had been
published, it was always listed as public.

This commit changes the behavior to take review request drafts into
consideration. If a draft exists, the "public" flag will be false.

As the output from the tests demonstrate, this bug has workflow
implications, as re-pushes to already published review requests may not
have been identified as "draft" as thus the user would not get prompted
to publish them.

One test change further demonstrates that the behavior on the server is
incorrect in the scenario where commits are re-pushed: if we perform a
"no-op" push, a new review request draft is created on the parent review
request. This is wrong and it has been reported in a new bug.