reviewboard: send reviewer info to client (bug 1177945); r=smacleod
An upcoming commit will allow the client to publish review requests
during push. We want the client to be in control of this operation. To
enable the client to make this decision, we need to give it data.

This commit sends reviewer info to the client.

The "reviewdata" mechanism has grown the ability to handle lists in
addition to plain strings. Lists are simply comma delimited strings of
URL quoted values. It's worth noting that ',' is URL encoded to '%2C'
and all values should be URL encoded so looking for ',' in the value
shouldn't have any false positives.

A new "listreviewdata" capability is now advertised as required to
ensure clients are capable of receiving lists in "reviewdata" fields.