reviewboard: support for publishing review requests (bug 1177945); r=smacleod
Having to visit a URL to publish review requests is annoying and wastes
developer time. Many developers have complained about this workflow.

This commit introduces support for publishing review requests when they
are submitted.

The server gains a wire protocol command for publishing review requests.
It advertises the presence of that command in the capabilities list.

Building on top of support for receiving reviewer information for
submitted review requests, the client now examines whether all submitted
review requests have reviewers attached. If so, it will prompt on
whether to publish.

I feel that people want the ability to auto publish during `hg push`.
This commit stops short of that and still requires a prompt. I have
concerns that there are enough bugs with commit mapping that have yet to
be worked out and that fully automatic publishing could result in too
much data loss. That is why auto publishing is not yet implemented.
However, depending on how well this step-in-the-right-direction feature
is received, we should consider adding it later if there is still
demand.