reviewboard: prevent reviews with public changesets (bug 1178957); r=smacleod
Before this commit, if the working copy was public or the user
explicitly requested review of public changesets, we might have
attempted to review public changesets. Coupled with the changes to
rewrite commits during push, and Mercurial would abort when trying to
mutate public changesets.

This commit does 2 things. First, it prints a message for each public
changeset that is dropped from the request to review. Second, it aborts
when there are no non-public changesets left for review and a review
cannot be conducted. Users now have a better understanding of what is
going on under the covers and they won't get a weird error when trying
to push public changesets.