deploy: support for creating review repositories
`deploy mozreview-create-repo` can now be used to create new review
repositories.

For the most part, "it just works." The one observed deficiency is that
the bug tracker type of the created repository is not set to "bugzilla."
This is because the Review Board web API does not appear to expose this
field as a primary field nor allows setting extra_data parameters via
the create() API (like review requests supports). This is a minor
deficiency and doesn't impact things in production at this time.