serverlog: parse hgweb commands out of URL
Previously, the repository in the log contained a path component. This
made aggregating all traffic belonging to a single repository difficult,
as there were multiple keys for a repository.

We now do some basic URL parsing and remove hgweb commands from URLs and
attribute the request to the proper repository. There is a slight chance
we may get a few false positives. We'll have to deploy this and see.

Ideally, we should just change serverlog to emit the full relative path
to the repository to make this problem go away. But that requires a
bunch of work to test different hgweb.wsgi files.