scripts: remote copy S3 object to reset lifecycle expiration (bug 1182739); r=bkero
We have a lifecycle policy on the S3 buckets for bundles that expires
objects after a few days (currently 3). If a repository is inactive for
a few days, we wouldn't upload bundles and the utilized S3 objects would
be expired.

This commit changes the behavior when the S3 object exists to perform a
remote copy, which will reset the object's mtime, thus resetting the
object expiration timer.