reviewboard: filter remote output on local ui instance (bug 1184599); r=smacleod
Mercurial 3.5 and/or bundle2 changed where the "remote: " output is
sent. Before, it was going to remote.ui. Now, it appears to be going to
repo.ui.

This commit adds our filtering wrapping to both ui instances.

In the not so distant future, we should leverage our new capabilities
foo or a bundle2 capability and remove this feature from the server. For
now, working around seems easier.