hgmo: link to Treeherder when possible (bug 1185023); r=bkero
Linking to Treeherder from the changeset pages is such an obviously
useful thing to do.

The easiest way to do this is for the repo's hgrc to define which
Treeherder repository it is associated with. This is much easier than
trying to parse the request's URL into a repository and mapping that
to Treeherder.

To make the template simpler, I also added a {pushhead} keyword to the
pushlog extension.